# src/project_forge/template/frontend/scripts/openapi/contracts/auth-sample.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Stable machine-readable error code",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "description": "Safe human-readable error message",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "description": "Login email",
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Plaintext password",
            "format": "password",
            "maxLength": 200,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "description": "Authenticated email",
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "expiresAt": {
            "description": "Session expiry",
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "userId": {
            "description": "Authenticated user ID",
            "format": "uuid",
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "email",
          "expiresAt"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "SignupRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "description": "Login email",
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Plaintext password",
            "format": "password",
            "maxLength": 200,
            "minLength": 12,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "workspaceName": {
            "description": "Initial workspace",
            "maxLength": 120,
            "minLength": 1,
            "title": "Workspacename",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "workspaceName"
        ],
        "title": "SignupRequest",
        "type": "object"
      },
      "WorkspaceListResponse": {
        "additionalProperties": false,
        "properties": {
          "workspaces": {
            "description": "Accessible workspaces",
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponse"
            },
            "title": "Workspaces",
            "type": "array"
          }
        },
        "required": [
          "workspaces"
        ],
        "title": "WorkspaceListResponse",
        "type": "object"
      },
      "WorkspaceRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "description": "Workspace name",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "WorkspaceRequest",
        "type": "object"
      },
      "WorkspaceResponse": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "description": "Workspace creation time",
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "name": {
            "description": "Workspace name",
            "title": "Name",
            "type": "string"
          },
          "workspaceId": {
            "description": "Workspace ID",
            "format": "uuid",
            "title": "Workspaceid",
            "type": "string"
          }
        },
        "required": [
          "workspaceId",
          "name",
          "createdAt"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Auth Matrix API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Origin denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication rate limit exceeded",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the fixed window resets",
                "schema": {
                  "minimum": 1.0,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Origin, CSRF, or authorization denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "getSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/signup": {
      "post": {
        "operationId": "signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Origin denied or signup disabled"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Email already exists"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication rate limit exceeded",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the fixed window resets",
                "schema": {
                  "minimum": 1.0,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Signup",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/workspaces": {
      "get": {
        "operationId": "listWorkspaces",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "List Workspaces",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "createWorkspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Origin, CSRF, or authorization denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Create Workspace",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}